improve timer #5
