+++ components/page.tsx
@@ -1,0 +1,40 @@
+import styled from 'styled-components'
+export { Container, Main, Footer, Logo }
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  min-height: 100vh;
+  padding: 0 0.5rem;
+`
+
+const Main = styled.main`
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 5rem 0;
+`
+const Footer = styled.footer`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100px;
+  border-top: 1px solid #eaeaea;
+
+  a {
+    display: flex;
+    flex-grow: 1;
+    align-items: center;
+    justify-content: center;
+  }
+`
+const Logo = styled.span`
+  height: 1em;
+  margin-left: 0.5rem;
+`
