+++ types/type.tsx
@@ -1,0 +1,4 @@
+export type { Pos, Values }
+
+type Pos = { x: number; y: number }
+type Values = { x: number; y: number; value: number }
